requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
chromadb
sentence-transformers
python-telegram-bot
openai
pdfplumber
